+++ changelog.md
@@ -3,7 +3,8 @@
 #### v1.8.0
 
 * `[sortutil]` Added case insensitive strings sorting
-* `[sliceutil]` Added deduplication function
+* `[sliceutil]` Added Deduplicate function
+* `[strutil]` Added ReplaceAll function
 
 #### v1.7.8